Rework GC
  * All types, except bool, Nil and Elipsis are now stored as a pointer inside
a Types::Value structure. This allows to easily have references to values when
dealing with closures (i.e. an int will not go out of scope because the outer
part of the closure is no longer in scope)
  * In consequence, refcounting is now separate from Types::Value, and relies on
a singleton GC class. Types::Value instances can now be freely exchanged around
without having to manually track references
  * All tests run fine and valgrind is happy. Next: continuing with variables!