Expressions now return a Var instead of a Value
  * A Var is a variant that can hold either a Types::Value, a Types::Value* or
a VarError. A Types::Value is an rvalue (not bound to a name), a Types::Value*
is an lvalue bound to a name, and a VarError is an invalid access (e.g. using
the dot operator or the right angle brackets on a number).
  * VarError is useful when we have to deal with the weird semantics of Lua
when dealing with multiple affectations on a single line and one of the
access on lhs of the = operator results in an error.