Fix several problems
  * Fix a crash caused by function returns of referenced values incorrectly
deleting the returned value if the local value that held it was the last
reference to it. This actually fixes a deeper problem related to visitVar_.
Now, if the context of the visit is not a visit in a varlist, visitVar_ will
always treat existing names as variables on the stack (i.e. put an rvalue in
the resulting Var, instead of a pointer), which will avoid the situation where
accessing a referenced value would crash the interpreter even if there were
other references to it.
  * Fixed the detection of the elipsis in expressions, it now properly uses
only the last three characters instead of the last four ones.
  * Affectations non properly update the reference count of values, removing
a reference to the old value before affectation and adding a new one after
affectation on the new value.
  * Fixed an issue where Function were not properly garbage collected resulting
in an immediate crash and/or several memory leaks.
  * Fixed visitPrefixexp not actually following its eventual nameAndArgs
attributes.
  * Fixed the adjustment of return values of function when multiple values are
returned. Previously, the next variable would hold a second copy of the first
returned value, instead of immediately skipping to the second. This would raise
exceptions when the second returned was a value, as it would usually hold
a non callable object.