Improved elipsis
  * Elipsis now stores copies of the input values instead of references, which
is more in line with what happens in Lua and avoids memory leaks. It also
unifies the content of an Elipsis with the return type of the function, as
both now store an array of Var / Values instead of an array of hard Vars and
an array of pointers to Var, which allows factorization when processing both
of them for adjustement.