Fix nil sometimes changing value
  * This was caused by the dot and subscript operators on tables appearing on
the left hand side of '='. In such cases, the use of references in the return
type of the dot and subscript methods of Table would sometimes cause a reference
to the static value nil to be returned, and subsequently changed. Now, the
returned value is copied in the context of Var_Context::OTHER, and a copy of
nil is inserted in the table at the requested field and returned in the context
of Var_Context::VARLIST.