Call C++ functions from Lua
  * The Environment class is similar to lua_State in C, and can be used to
register functions as global variables
  * FunctionAbstraction and CurriedFunction are used to represent, respectively,
an abstracted form of a function which is in a curried form and can be
progressively composed by applying values to it and finally called, and a
function that is progressively reduced until it becomes a no argument function
that can be called. This mirrors the process by which Ramsay embed OCaml
functions in LuaML.
  * FunctionAbstractionBuilderAbstraction is a helper class to build a new
instance of a C++ function in the interpreter.
  * meta_types.h stores several metafunctions that are used to parse complex
types, such as function types (extract parameters, reduce signatures), or
build complex types (building a variant back from a list of type that has been
abstracted away into an ArgsList for example)
  * Converter is a new class that stores the conversions functions the interpreter
can use to perform the conversion from Lua values to C++ values when needed.
This is similar to the map / mapf types used by Ramsay to implement its
conversion functions in LuaML.